move legend test to barGauge_new.tests.js
